requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2.1
django-filter==25.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
importlib_resources==6.5.2
inflection==0.5.1
jalali_core==1.0.0
jdatetime==5.2.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
Markdown==3.8
packaging==25.0
phonenumbers==9.0.5
pillow==11.2.1
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
swagger-spec-validator==3.0.4
typing_extensions==4.13.2
uritemplate==4.1.1
whitenoise==6.9.0
